refactor: Remove references to elf::secnames from resolution
